add tests

+++ src/main/java/kz/beeline/aprudnikov/service/TestTaskEntityFileManager.java
@@ -98,13 +98,14 @@ public class TestTaskEntityFileManager {
     public void deleteFolder(String directory) {
         File file = new File(directory);
         File[] files = file.listFiles();
-        assert files != null;
-        for (File f:files) {
-            if (f.isFile() && f.exists()) {
-                boolean deleteResult = f.delete();
-                System.out.println(f.getName() + " deleted: " + deleteResult);
-            } else {
-                System.out.println("can not delete a file due to open or error");
+        if (files != null) {
+            for (File f:files) {
+                if (f.isFile() && f.exists()) {
+                    boolean deleteResult = f.delete();
+                    // System.out.println(f.getName() + " deleted: " + deleteResult);
+                } else {
+                    System.out.println("can not delete a file due to open or error");
+                }
             }
         }
     }
